update sql
修改sql文件

+++ database_sql/power.sql
@@ -30,7 +30,7 @@ CREATE TABLE `unit_capping`  (
   `unit_capping_id` int(10) NOT NULL AUTO_INCREMENT COMMENT 'capping动作的唯一主键',
   `unit_capping_name` varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL DEFAULT '' COMMENT 'capping动作的名称或者描述',
   `func_unit_id` int(10) NOT NULL COMMENT '对应功能部件的id',
-  `capping_value` int(5) NOT NULL COMMENT 'capping的值',
+  `capping_value` float(10) NOT NULL COMMENT 'capping的值',
   `is_del` varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NOT NULL DEFAULT '0' COMMENT '是否删除(0:未删除；1:已删除)',
   `create_date` timestamp(0) NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间，记录capping动作的时间',
   PRIMARY KEY (`unit_capping_id`) USING BTREE
@@ -41,9 +41,9 @@ CREATE TABLE `unit_power`  (
   `unit_power_id` int(10) NOT NULL AUTO_INCREMENT COMMENT '部件实时功耗的唯一id',
   `unit_power_name` varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL DEFAULT '' COMMENT '部件实时功耗名称或者描述',
   `func_unit_id` int(10) NOT NULL COMMENT '对应功能部件的id',
-  `unit_power` int(5) NOT NULL COMMENT '功能部件的功耗',
-  `unit_usage` int(5) NOT NULL DEFAULT 0 COMMENT '功能部件的利用率',
-  `unit_temperature` int(5) NOT NULL DEFAULT 0 COMMENT '功能部件的温度',
+  `unit_power` float(10) NOT NULL COMMENT '功能部件的功耗',
+  `unit_usage` float(10) NOT NULL DEFAULT 0.0 COMMENT '功能部件的利用率',
+  `unit_temperature` float(10) NOT NULL DEFAULT 0.0 COMMENT '功能部件的温度',
   `is_del` varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NOT NULL DEFAULT '0' COMMENT '是否删除(0:未删除；1:已删除)',
   `create_date` timestamp(0) NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
   PRIMARY KEY (`unit_power_id`) USING BTREE
